Updated sample for V2 client
