+++ modules/FilterModule/Dockerfile.amd64.debug
@@ -1,16 +1,12 @@
-FROM ubuntu:xenial
+FROM ubuntu:20.04
 
 WORKDIR /app
 
-RUN apt-get update && \
-    apt-get install -y --no-install-recommends libcurl4-openssl-dev python3-pip libboost-python1.58-dev libpython3-dev && \
-    rm -rf /var/lib/apt/lists/* 
+RUN apt-get update
+RUN apt-get install -y python3-pip
 
-RUN pip3 install --upgrade pip
-RUN pip install setuptools
-RUN pip install ptvsd==4.1.3
 COPY requirements.txt ./
-RUN pip install -r requirements.txt
+RUN pip3 install -r requirements.txt
 
 COPY . .
