+++ modules/FilterModule/Dockerfile.arm32v7
@@ -1,19 +1,16 @@
-FROM resin/raspberrypi3-debian:stretch
+FROM ubuntu:20.04
 
 WORKDIR /app
 
-RUN apt-get update && \
-    apt-get install -y --no-install-recommends libboost-python1.62.0 python3-pip libpython3-dev && \
-    rm -rf /var/lib/apt/lists/* 
+RUN apt-get update
+RUN apt-get install -y python3-pip
 
-RUN pip3 install --upgrade pip 
-RUN pip install --upgrade setuptools 
 COPY requirements.txt ./
-RUN pip install -r requirements.txt
+RUN pip3 install -r requirements.txt
 
 COPY . .
 
 RUN useradd -ms /bin/bash moduleuser
 USER moduleuser
 
-ENTRYPOINT [ "python3", "-u", "./main.py" ]
+CMD [ "python3", "-u", "./main.py" ]
