+++ modules/FilterModule/requirements.txt
@@ -1,1 +1,1 @@
-azure-iothub-device-client~=1.4.3
+azure-iot-device
